+++ src/Firebase/Exception/MessagingException.php
@@ -21,46 +21,62 @@ class MessagingException extends \RuntimeException implements FirebaseException
      */
     private $response;
 
+    /**
+     * @var array
+     */
+    private $errors = [];
+
     public static function fromRequestException(RequestException $e): self
     {
-        $message = $e->getMessage();
+        $errors = [];
+        $reasonPhrase = null;
 
-        switch ($code = $e->getCode()) {
+        if ($response = $e->getResponse()) {
+            $errors = self::getErrorsFromResponse($response);
+            $reasonPhrase = $response->getReasonPhrase();
+        }
+
+        $code = (int) ($errors['error']['code'] ?? $e->getCode());
+        $message = $errors['error']['message'] ?? $reasonPhrase;
+
+        switch ($code) {
             case 400:
-                $error = new InvalidMessage($message, $code);
+                $error = new InvalidMessage($message ?: 'Invalid message', $code);
                 break;
             case 401:
             case 403:
-                $error = new AuthenticationError($message, $code);
+                $error = new AuthenticationError($message ?: 'Authentication error', $code);
                 break;
             case 404:
-                $error = new NotFound($message, $code);
+                $error = new NotFound($message ?: 'Not found', $code);
                 break;
             case 500:
-                $error = new ServerError($message, $code);
+                $error = new ServerError($message ?: 'Server error', $code);
                 break;
             case 503:
-                $error = new ServerUnavailable($message, $code);
+                $error = new ServerUnavailable($message ?: 'Server unavailable', $code);
                 break;
             default:
-                $error = new UnknownError($message, $code);
+                $error = new UnknownError($message ?: 'Unknown error', $code);
                 break;
         }
 
-        if ($e->hasResponse()) {
-            $error = $error->withResponse($e->getResponse());
-        }
-
-        return $error;
+        return $error
+            ->withResponse($response)
+            ->withErrors($errors);
     }
 
     public function errors(): array
     {
-        if (!$this->response || !JSON::isValid($body = $this->response->getBody())) {
-            return [];
-        }
+        return $this->errors;
+    }
 
-        return JSON::decode($body, true);
+    public function withErrors(array $errors = [])
+    {
+        $e = new static($this->getMessage(), $this->getCode());
+        $e->errors = $errors;
+
+        return $e;
     }
 
     /**
@@ -77,5 +93,14 @@ class MessagingException extends \RuntimeException implements FirebaseException
         $e->response = $response;
 
         return $e;
+    }
+
+    private static function getErrorsFromResponse(ResponseInterface $response): array
+    {
+        try {
+            return JSON::decode((string) $response->getBody(), true);
+        } catch (\InvalidArgumentException $e) {
+            return [];
+        }
     }
 }